cleaned up two places checking for null, but should be checking for undefined. (#97)
The Location.create method returns a valid location object or undefined, it never returns null.
The fix I checked in earlier, I discovered has to do with JSON. stringify converts an array's undefined elements to null. so we need to check for null on the webview side for elements in the location array.