Add functionality to inject analysis on startup (#522)
* Add functionality to inject analysis on startup

If the `sarif-viewer.githubCodeScanningInitialAlert` setting has a
value, use the contents of the value as the alert to apply. This is
meant as a way to automatically apply fixes from a codescanning alert.

Additionally, if this alert has any fixes in it, they are automatically
applied. We also need to handle some slight differences in the SARIF
format. The fix objects that we are receiving are non-standard. We
expect that for each fix, there is a `diff` in its property bag that
contains a git-patch formatted version of the diff to apply.

Lastly, this adds some simple telemetry. It collects that the
autofix was applied, whether or not it was successful, and (if
unsuccessful) a failure message.

* Restructure `activateGithubAnalyses`

* Add retries for getting the remote fetch url

On codespaces, there is a race condition where sometimes the extension
starts before the filesystem is fully activated. When this happens, it
is not possible to get any of the remotes from the git api.

The solution is to perform some retries when getting the remotes.

* Add more logging to the output channel

No functional changes, just help with debugging and information.

* Address comments from review

* Bump version